refactor(StringCalculator): 코드 리뷰 반영
- Expression 패키지 분할
- 유틸성 객체 object 변경
- 매직 넘버 상수화
- 파일 끝 개행 반영
