+++ src/main/kotlin/stringcalculator/expression/Expression.kt
@@ -1,0 +1,6 @@
+package stringcalculator.expression
+
+data class Expression(
+    val delimiters: List<String>,
+    val input: String,
+)
